requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==1.0.0
cryptography==41.0.3
pyotp==2.9.0
bcrypt==4.0.1
python-dotenv==0.19.0
Flask-Migrate==4.0.4
email-validator==1.1.3
Flask-Talisman==1.0.0
Flask-SeaSurf==1.1.1
Werkzeug==2.0.3
SQLAlchemy==1.4.23
pytest==7.4.0
pyOpenSSL==23.2.0
